requirements.txt
flask==3.0.2
python-quickbooks==0.9.12
intuit-oauth==1.2.6